+++ src/lib/types/index.d.ts
@@ -1,18 +1,6 @@
 // https://www.typescriptlang.org/docs/handbook/jsdoc-supported-types.html
-export type Library = {
-  id: string;
-  name: string;
-  created: Date;
-  modified: Date;
-}
-
-export type LibraryPath = {
-  id: string;
-  libraryId: string;
-  path: string;
-  created: Date;
-  modified: Date;
-}
+export * from "./job"
+export * from "./library"
 
 export type Video = {
   id: string;
