+++ src/lib/types/library.d.ts
@@ -1,0 +1,14 @@
+export type Library = {
+  id: string;
+  name: string;
+  created: Date;
+  modified: Date;
+}
+
+export type LibraryPath = {
+  id: string;
+  libraryId: string;
+  path: string;
+  created: Date;
+  modified: Date;
+}
